+++ feature/notice/src/main/java/com/wap/wapp/feature/notice/NoticeFragment.kt
@@ -1,7 +1,6 @@
 package com.wap.wapp.feature.notice
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -38,9 +37,8 @@ class NoticeFragment : Fragment() {
 
         composeView.setContent {
             val events by viewModel.events.collectAsState()
-            Log.d("test", "NoticeFragment : $events")
-
             val dateAndDayOfWeek = dateUtil.generateNowDateAndDay()
+
             WappTheme {
                 NoticeScreen(events, dateAndDayOfWeek)
             }
